Injecao de dependencias e tratamento

+++ code/app_locadora_carros/app/Http/Controllers/MarcaController.php
@@ -5,15 +5,22 @@ namespace App\Http\Controllers;
 use App\Models\Marca;
 use Illuminate\Http\Request;
 
+
 class MarcaController extends Controller
 {
+    protected $marca;
+    public function __construct(Marca $marca)
+    {
+        $this->marca=$marca;
+    }
     /**php artisan make:model --migration --controller --resource Marca
 
      * Display a listing of the resource.
      */
     public function index()
     {
-        $marcas = Marca::all();
+        //$marcas = Marca::all();
+        $marcas=$this->marca->all();
         return $marcas;
     }
 
@@ -30,16 +37,20 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca=Marca::create($request->all());
-        
+        //$marca=Marca::create($request->all());
+        $marca= $this->marca->create($request->all());
         return $marca;
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Marca $marca)
+    public function show(int $id)
     {
+        $marca= $this->marca->find($id);
+        if($marca===null){
+            return ['erro'=>'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -54,21 +65,36 @@ class MarcaController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request,  int $id)
     {
-        print_r($request->all()); //dados atualizados
-        echo '<hr>';
-        print_r($marca->getAttributes()); //dados antigos
+        // print_r($request->all()); //dados atualizados
+        // echo '<hr>';
+        // print_r($marca->getAttributes()); //dados antigos
 
+
+        //$marca->update($request->all());
+
+        $marca=$this->marca->find($id);
+
+        if($marca===null){
+            return ['erro'=>'Impossivel realizar a atualização.'];
+        }
         $marca->update($request->all());
+        return $marca;
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Marca $marca)
+    public function destroy(int $id)
     {
+        $marca= $this->marca->find($id);
+
+        if($marca===null){
+            return ['erro'=>'Impossivel realizar a exclusao.'];
+        }
         $marca->delete();
+        // $marca->delete();
         return ['msg'=>'A marca foi removida com sucesso!'];
     }
 }
